+++ policy-engine/src/main/java/com/walmartlabs/concord/policyengine/ForkDepthPolicy.java
@@ -1,0 +1,46 @@
+package com.walmartlabs.concord.policyengine;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import java.util.Collections;
+import java.util.concurrent.Callable;
+
+public class ForkDepthPolicy {
+
+    private final ForkDepthRule rule;
+
+    public ForkDepthPolicy(ForkDepthRule rule) {
+        this.rule = rule;
+    }
+
+    public CheckResult<ForkDepthRule, Integer> check(Callable<Integer> c) throws Exception {
+        if (rule == null) {
+            return new CheckResult<>();
+        }
+
+        int depth = c.call();
+        if (depth >= rule.getMax()) {
+            return new CheckResult<>(Collections.emptyList(),
+                    Collections.singletonList(new CheckResult.Item<>(rule, depth)));
+        }
+        return new CheckResult<>();
+    }
+}